requirements.txt
# AI Pharma Lead Generation Platform
# Dependencies

langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
openai>=1.0.0
google-search-results>=2.4.2
pydantic>=2.0.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
